refactor: fetchClient, getHeaders 폴더 구조 수정 및 fetchClient 로직 수정
- undefined를 반환하는 문제를 가진 fetchClient함수의 try, catch 임시 삭제 및 추후 논의
- fetchClient 반환 데이터 수정
- header 타입 수정
