+++ frontend/src/apis/_common/getHeaders.ts
@@ -2,7 +2,7 @@ import { getCookie } from '@utils/cookies';
 
 import { COOKIE_KEYS } from '@constants/cookies';
 
-export default function getHeaders(): Record<string, string> {
+export default function getHeaders(): HeadersInit {
   const headers = { 'Content-type': 'application/json' };
   const token = getCookie(COOKIE_KEYS.token);
